% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bRushVerbose.R
\name{bRushVerbose}
\alias{bRushVerbose}
\title{Verbose console output}
\usage{
bRushVerbose(response)
}
\arguments{
\item{response}{http response object.}
}
\description{
For debugging, etc., outputs informations after each API request
}
