% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkErrors.R
\name{checkErrors}
\alias{checkErrors}
\title{Check for Errors}
\usage{
checkErrors(status)
}
\arguments{
\item{status}{http status passed by process request.}
}
\description{
Used by processRequest to stop if there are http errors.
}
