% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAdmin.R
\name{createAdmin}
\alias{createAdmin}
\title{Create a content migration.}
\usage{
createAdmin(
  accountID,
  userID,
  roleID,
  send_confirmation = FALSE,
  server = "test",
  ...
)
}
\arguments{
\item{accountID}{Account ID of context to import into.}

\item{userID}{User ID of individual to make admin.}

\item{roleID}{ID number of the role.}

\item{send_confirmation}{Boolean to send notification}

\item{server}{Test, beta, prodcution, OR alternative name in R.environ OR url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Create a content migration with migrator type and a settings list. Does not
currently support file uploads.  Limited testing conducted on various migrators.
}
