% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAssignment.R
\name{createAssignment}
\alias{createAssignment}
\title{Create a new assignment}
\usage{
createAssignment(courseID, name, description = "", points_possible = "",
  type = "online", online_type = c("online_upload", "online_text_entry"),
  allowed_extensions = c("doc", "docx", "pdf", "ppt", "pptx", "rtf"),
  grading_type = "points", assignment_group_id = "",
  grading_standard_id = "", group_category_id = "", due_at = "",
  lock_at = "", unlock_at = "", published = FALSE,
  turnitin_enabled = FALSE, turnitin_settings = NULL,
  peer_reviews = FALSE, automatic_peer_reviews = FALSE,
  notify_of_update = FALSE, muted = FALSE,
  grade_group_students_individual = FALSE, ext_tool_url = NULL,
  ext_tool_new_tab = FALSE, server = "test", ...)
}
\arguments{
\item{courseID}{Course ID to create the assignment in.}

\item{name}{Character of the name of the assignment}

\item{description}{Character description of the assignments, supports HTML.}

\item{points_possible}{Integer of the maximum number of points possible}

\item{type}{One character of assignment type ("online","none", "on_paper", "online_quiz", "discussion_topic", "external_tool")}

\item{online_type}{If type is "online", vector of online types ("online_upload", "online_text_entry", "online_url", "media_recording")}

\item{allowed_extensions}{Character of the extensions allowed for file submissions for "online_upload"}

\item{grading_type}{Character of grading type options (pass_fail, percent, letter_grade, gpa_scale, points)}

\item{assignment_group_id}{Integer of the assignment group to put assignment in.}

\item{grading_standard_id}{ID of the grading standard to be used}

\item{group_category_id}{ID of group to assign discussion to.}

\item{due_at}{POSIXct object of date and time of due date}

\item{lock_at}{POSIXct object of date and time of lock date}

\item{unlock_at}{POSIXct object of date and time of unlock date}

\item{published}{Boolean to specify if the assignment is published for availability.}

\item{turnitin_enabled}{Boolean to specify whether to use the Turnitin API.}

\item{turnitin_settings}{Named list of turnitin settings.}

\item{peer_reviews}{Boolean to enable peer reviews}

\item{automatic_peer_reviews}{Boolean to enable automatically assign peer reviewers}

\item{notify_of_update}{Boolean to notify students that assignment content has changed.}

\item{muted}{Boolean to mute assignment or not}

\item{ext_tool_url}{Character of external tool url if type is "external_tool".}

\item{ext_tool_new_tab}{Boolean of whether ext tool opens in new tab.}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}

\item{external_tool_tag_attributes}{Named list od external tool attributes.}

\item{grade_group_students_individually}{Boolean to say group students will be graded individually.}
}
\description{
Create a new assignment for the provided course.
}
