% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createContentExport.R
\name{createContentExport}
\alias{createContentExport}
\title{Create a content export}
\usage{
createContentExport(
  ID,
  export_type,
  type = "course",
  skip_notifications = TRUE,
  select = NULL,
  server = "test",
  ...
)
}
\arguments{
\item{ID}{ID of context to import into.}

\item{export_type}{Export type - common_cartridge, qti, zip}

\item{type}{Type - course, group, or user}

\item{skip_notifications}{Notify the user of the export creation}

\item{select}{Named list of object types and the IDs to include}

\item{server}{Test, beta, prodcution, OR alternative name in R.environ OR url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Create a new content export.
}
