% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createConversation.R
\name{createConversation}
\alias{createConversation}
\title{Create a new conversation}
\usage{
createConversation(url, userID, subject = NULL, body = NULL, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{userID}{User id of person to message}

\item{subject}{Character of conversation subject}

\item{body}{Character of body content}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Create a single recipient conversation
}
