% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFile.R
\name{createFile}
\alias{createFile}
\title{Upload a course file
Upload}
\usage{
createFile(url, courseID, name = "", filelocal = "", path = "", ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{courseID}{Course id to create upload the file to}

\item{name}{The name of the file upon upload.}

\item{filelocal}{The path to the file.}

\item{path}{Path of the folder in course files.}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Upload a course file
Upload
}
