% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMigration.R
\name{createMigration}
\alias{createMigration}
\title{Create a content migration.}
\usage{
createMigration(url, ID, source_course_id, type = "course",
  migration_type = "course_copy_importer", ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{ID}{ID of context to import into.}

\item{type}{String of type - course, account, group, or user}

\item{migration_type}{String of migrator type}

\item{...}{Optional page options to pass to processRequest}

\item{settings}{List of settings for the migrator}
}
\description{
Create a content migration with migrator type and a settings list.
}
