% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutcome.R
\name{createOutcome}
\alias{createOutcome}
\title{Create a new outcome in an outcome group}
\usage{
createOutcome(
  ID,
  groupID,
  type = "account",
  title,
  display_name = "",
  description = "",
  mastery_points = NULL,
  ratings = NULL,
  calculation_method = "highest",
  calculation_int = NULL,
  vendor_guid = NULL,
  server = "test",
  ...
)
}
\arguments{
\item{ID}{Account ID of the outcome group to create the outcome in}

\item{groupID}{Outcome group id to create the outcome in}

\item{type}{Character of "course" or "account"}

\item{title}{Character of the title of the outcome}

\item{display_name}{Character of a friendly name for titles that are vague/numbered.}

\item{description}{Character of the description of the outcome}

\item{mastery_points}{Integer of the mastery threshold in the rubric criterion}

\item{ratings}{A list of named vectors with "description" & "points" defined for each criterion}

\item{calculation_method}{Character of the method of calculation: decaying_average, n_mastery, latest, highest}

\item{calculation_int}{Integer if method is decaying_average(percent higher rate recent count as) or n_mastery (n times mastery must be acheived)}

\item{vendor_guid}{A custom GUID for the learning standard.}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Create a new outcome in an outcome group.
}
