% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPage.R
\name{createPage}
\alias{createPage}
\title{Create a new wiki page}
\usage{
createPage(url, ID, title = NULL, body = NULL, edit = "teachers",
  notify = FALSE, publish = TRUE, front = FALSE, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{ID}{Course or group id to create the page in}

\item{title}{String of page title.}

\item{body}{String of body content.}

\item{edit}{String of roles that can edit page ("teacher","student","members,"public")}

\item{notify}{Boolean to specify whether users should be notified of creations}

\item{publish}{Boolean to publish the page after creating.}

\item{front}{Boolean to set page as the front page}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Create a new wiki page for the provided course or group.
}
