% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSection.R
\name{createSection}
\alias{createSection}
\title{Create a new section}
\usage{
createSection(
  courseID,
  name,
  sis_section_id = "",
  start_at = "",
  end_at = "",
  restrict = TRUE,
  server = "test",
  ...
)
}
\arguments{
\item{courseID}{Course ID to create the section in.}

\item{name}{Character of the name of the section}

\item{sis_section_id}{Character of the sis id of the section}

\item{start_at}{Section start date in ISO8601 format, e.g. 2011-01-01T01:00Z}

\item{end_at}{Section end date in ISO8601 format, e.g. 2011-01-01T01:00Z}

\item{restrict}{Boolean, set to true to restrict user enrollments to the start and end dates of the section.}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Create a new section for the provided course.
}
