% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteEnrollment.R
\name{deleteEnrollment}
\alias{deleteEnrollment}
\title{Delete or inactivate an enrollment}
\usage{
deleteEnrollment(courseID, userID, task = "delete", server = "test", ...)
}
\arguments{
\item{courseID}{Course ID of enrollment}

\item{userID}{User ID of enrollment}

\item{task}{One of "conclude", "delete", "inactivate", "deactivate"}

\item{server}{Test, beta, prodcution, OR alternative name in R.environ OR url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get user details for the provided user id. NOTE: Returns a list.
}
