% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editAssignment.R
\name{editAssignment}
\alias{editAssignment}
\title{Edit an assignment}
\usage{
editAssignment(url, courseID, assignID, name = NULL, description = NULL,
  points_possible = NULL, type = NULL, online_type = NULL,
  allowed_extensions = NULL, grading_type = NULL,
  assignment_group_id = NULL, grading_standard_id = NULL,
  group_category_id = NULL, due_at = NULL, lock_at = NULL,
  unlock_at = NULL, published = NULL, turnitin_enabled = NULL,
  peer_reviews = NULL, automatic_peer_reviews = NULL,
  notify_of_update = NULL, muted = NULL,
  grade_group_students_individual = NULL, ext_tool_url = NULL,
  ext_tool_new_tab = NULL, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation.}

\item{courseID}{Course ID to create the assignment in.}

\item{assignID}{Assignment ID of the assignment to edit.}

\item{name}{Character of the name of the assignment}

\item{description}{Character description of the assignments, supports HTML.}

\item{points_possible}{Integer of the maximum number of points possible}

\item{type}{One character of assignment type ("online","none", "on_paper", "online_quiz", "discussion_topic", "external_tool")}

\item{online_type}{If type is "online", vector of online types ("online_upload", "online_text_entry", "online_url", "media_recording")}

\item{allowed_extensions}{Character of the extensions allowed for file submissions for "online_upload"}

\item{grading_type}{Character of grading type options (pass_fail, percent, letter_grade, gpa_scale, points)}

\item{assignment_group_id}{Integer of the assignment group to put assignment in.}

\item{grading_standard_id}{Integer id of the grading standard used.}

\item{group_category_id}{ID of group to assign discussion to.}

\item{due_at}{POSIXct object of date and time of due date}

\item{lock_at}{POSIXct object of date and time of lock date}

\item{unlock_at}{POSIXct object of date and time of unlock date}

\item{published}{Boolean to specify if the assignment is published for availability.}

\item{turnitin_enabled}{Boolean to specify whether to use the Turnitin API.}

\item{peer_reviews}{Boolean to enable peer reviews}

\item{automatic_peer_reviews}{Boolean to enable automatically assign peer reviewers}

\item{notify_of_update}{Boolean to notify students that assignment content has changed.}

\item{muted}{Boolean to mute assignment or not}

\item{ext_tool_url}{Character of external tool url if type is "external_tool".}

\item{ext_tool_new_tab}{Boolean of whether ext tool opens in new tab.}

\item{...}{Optional page options to pass to processRequest}

\item{turnitin_settings}{Named list of turnitin settings.}

\item{external_tool_tag_attributes}{Named list od external tool attributes.}

\item{grade_group_students_individually}{Boolean to say group students will be graded individually.}
}
\description{
Edit an assignment of the provided ID for the provided course.
}
