% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editDiscussionTopic.R
\name{editDiscussionTopic}
\alias{editDiscussionTopic}
\title{Edit a discussion topic}
\usage{
editDiscussionTopic(
  courseID,
  topicID,
  title = NULL,
  message = NULL,
  discussion_type = NULL,
  published = NULL,
  podcast_enabled = NULL,
  podcast_has_student_posts = NULL,
  require_initial_post = NULL,
  is_announcement = NULL,
  pinned = NULL,
  allow_rating = NULL,
  only_graders_can_rate = NULL,
  sort_by_rating = NULL,
  delayed_post_at = NULL,
  lock_at = NULL,
  position_after = NULL,
  group_category_id = NULL,
  points_possible = NULL,
  peer_reviews = NULL,
  automatic_peer_reviews = NULL,
  grading_type = NULL,
  grading_standard_id = NULL,
  due_at = NULL,
  unlock_at = NULL,
  server = "test",
  ...
)
}
\arguments{
\item{courseID}{Course or group ID to create the discussion in.}

\item{topicID}{ID for the discussion topic}

\item{title}{Character of the title of the assignment}

\item{message}{Character of the discussion message.}

\item{discussion_type}{Character of the discussion type ("side_comment" or "threaded")}

\item{published}{Boolean to specify if the discussion is published for availability.}

\item{podcast_enabled}{Boolean to enable podcast feed.}

\item{podcast_has_student_posts}{Boolean to include student replies in podcast feed.}

\item{require_initial_post}{Boolean to specify if user must post first to begin to reply.}

\item{is_announcement}{Boolean to specify if this is an announcment instead of a discussion.}

\item{pinned}{Boolean to list as a pinned discussion.}

\item{allow_rating}{Boolean to allow users to rate entries.}

\item{only_graders_can_rate}{Boolean to allow only graders to rate.}

\item{sort_by_rating}{Boolean to sort by rating.}

\item{delayed_post_at}{POSIXct object of date and time to publish discussion.}

\item{lock_at}{POSIXct object of date and time of lock date if an assignment}

\item{position_after}{ID of discussion topic to position this one after.}

\item{group_category_id}{ID of group to assign discussion to.}

\item{points_possible}{Integer of the maximum number of points possible}

\item{peer_reviews}{Boolean to enable peer reviews}

\item{automatic_peer_reviews}{Boolean to enable automatically assign peer reviewers}

\item{grading_type}{Character of grading type options (pass_fail, percent, letter_grade, gpa_scale, points)}

\item{grading_standard_id}{ID of the grading standard to be used}

\item{due_at}{POSIXct object of date and time of due date if an assignment}

\item{unlock_at}{POSIXct object of date and time of unlock date if an assignment}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest if an assignment}
}
\description{
Edit an exisitng discusson for the provided course or group.
}
