% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editGradebookColumnData.R
\name{editGradebookColumnData}
\alias{editGradebookColumnData}
\title{Create/edit a custom grade book column data entry for user in a course}
\usage{
editGradebookColumnData(url, courseID, columnID, userID, content, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{courseID}{Course ID with the custom column}

\item{columnID}{ID of of the column}

\item{userID}{ID of the user (student) to write column datum to}

\item{...}{Optional page options to pass to processRequest}

\item{include_hidden}{Character of the content to place in the column for user}
}
\description{
Create/edit a new custom grade book column entry for the provided user and course.
}
