% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollUser.R
\name{enrollUser}
\alias{enrollUser}
\title{Enroll a user in a section or course}
\usage{
enrollUser(url, ID, user_id, course_section_id = "",
  type = "StudentEnrollment", enrollment_state = "active",
  limit_privileges = TRUE, notify = "")
}
\arguments{
\item{url}{The base url of a Canvas installation.}

\item{ID}{Course ID for enrollment}

\item{user_id}{ID of the user to be enrolled}

\item{course_section_id}{ID of the course section to enroll student in}

\item{type}{Character of StudentEnrollment, TeacherEnrollment, TaEnrollment, ObserverEnrollment, DesignerEnrollment}

\item{enrollment_state}{Character of active, invited, inactive}

\item{limit_privileges}{Boolean, if true, the enrollment will only allow the user to see and interact with users enrolled in the section given by course_section_id}

\item{notify}{Boolean, if true, a notification will be sent to the enrolled user.}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Create a new section for the provided course.
}
