% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccountUsers.R
\name{getAccountUsers}
\alias{getAccountUsers}
\title{Get Users from an Account}
\usage{
getAccountUsers(url, accountID, search_term = NULL, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{accountID}{Integer of the account ID to find users for}

\item{search_term}{Filter results by partial name or full ID to match and return in the results list. Must be at least 3 characters.}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get the users from a specified account.
}
