% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdmins.R
\name{getAdmins}
\alias{getAdmins}
\title{Get a list of admins for an account}
\usage{
getAdmins(accountID, server = "test", ...)
}
\arguments{
\item{accountID}{Account ID to retereive admins of}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of roles for an account
}
