% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnalytics.R
\name{getAnalytics}
\alias{getAnalytics}
\title{Get account or course analytics}
\usage{
getAnalytics(ID, type = "activity", term = "current", course = FALSE,
  studentID = NULL, server = "test", ...)
}
\arguments{
\item{type}{The analytics type (e.g., activity, grades, statistics, assignments, communication, student_summaries)}

\item{term}{The term ID, "complete", or "current"(for account analytics only)}

\item{course}{Boolean to specify it the request is for a course.}

\item{studentID}{ID of a student for student specific requests (activity, assignment, communication)}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get account or course analytics
}
