% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssignment.R
\name{getAssignment}
\alias{getAssignment}
\title{Get the a single assignment}
\usage{
getAssignment(
  courseID,
  assignmentID,
  include = c("submission", "assignment_visibility", "overrides", "observed_users",
    "can_edit", "score_statistics"),
  all_dates = TRUE,
  server = "test",
  ...
)
}
\arguments{
\item{courseID}{Course ID of the outcome}

\item{assignmentID}{Vector of user IDs of the students}

\item{include}{Options to include in the results}

\item{all_dates}{Boolean to include results hidden from learning mastery gradebook}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get the a single assignment
}
