% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssignments.R
\name{getAssignments}
\alias{getAssignments}
\title{Get assignments from a course}
\usage{
getAssignments(
  courseID,
  search_term = NULL,
  bucket = NULL,
  server = "test",
  ...
)
}
\arguments{
\item{courseID}{Course ID to find assignments for}

\item{search_term}{Filter results by partial course name, code, or full ID to match and return in the results list. Must be at least 3 characters.}

\item{bucket}{Limit results to a named bucket ("past", "overdue", "undated", "ungraded", "upcoming", and "future")}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get assignments for the provided course. Most include parameters ("email",
"enrollments","locked","avatar_url","bio") for this request have all been 
enabled and can be subsetted out if not needed.  This excludes "test_student,"
which has been set to its own argument.
}
