% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContentExport.R
\name{getContentExport}
\alias{getContentExport}
\title{Show a single content export}
\usage{
getContentExport(ID, exportID, type = "course", server = "test", ...)
}
\arguments{
\item{ID}{Integer of the account ID to find sub-accounts for}

\item{exportID}{Integer of the specific content export to return}

\item{type}{String of type - course, account, group, or user}

\item{server}{Test, beta, prodcution, OR alternative name in R.environ OR url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
List content exports
}
