% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCourse.R
\name{getCourse}
\alias{getCourse}
\title{Get a single course}
\usage{
getCourse(courseID, server = "test", ...)
}
\arguments{
\item{courseID}{The base url of a Canvas installation}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get details of a single course.
}
