% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCourseAudit.R
\name{getCourseAudit}
\alias{getCourseAudit}
\title{Get a course audit object}
\usage{
getCourseAudit(
  courseID,
  start_time = NULL,
  end_time = NULL,
  server = "test",
  ...
)
}
\arguments{
\item{courseID}{Course or group ID of page}

\item{start_time}{Start date and in ISO8601 format, e.g. 2011-01-01T01:00Z}

\item{end_time}{End date and in ISO8601 format, e.g. 2011-01-01T01:00Z}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of pages in course or group.
}
