% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCourses.R
\name{getCourses}
\alias{getCourses}
\title{Get courses from an account}
\usage{
getCourses(accountID, termID = NULL, search_term = NULL,
  by_teachers = NULL, by_subaccounts = NULL, type = NULL,
  server = "test", ...)
}
\arguments{
\item{accountID}{Integer of the account ID to find sub-accounts for}

\item{termID}{Integer of enrollment term ID}

\item{search_term}{Filter results by partial course name, code, or full ID to match and return in the results list. Must be at least 3 characters.}

\item{by_teachers}{Vector of user ID integers.}

\item{by_subaccounts}{Vector of subaccount integers.}

\item{type}{Filter results to include course with at least one of specified string ("teacher","student","ta","observer","designer")}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get course details for the provided account. All include parameters ("syllabus_body”, 
“term”, “course_progress”, “storage_quota_used_mb”, “total_students”, “teachers) 
for this request have all been enabled and can be subsetted out if not needed.
}
