% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCourseUsers.R
\name{getCourseUsers}
\alias{getCourseUsers}
\title{Get users from a course}
\usage{
getCourseUsers(
  courseID,
  search_term = NULL,
  user_ids = NULL,
  test_student = FALSE,
  type = NULL,
  role_id = NULL,
  state = NULL,
  server = "test",
  ...
)
}
\arguments{
\item{courseID}{Course ID to find users for}

\item{search_term}{Filter results by partial course name, code, or full ID to match and return in the results list. Must be at least 3 characters.}

\item{user_ids}{Vector of user IDs to limit results by}

\item{test_student}{Boolean to include test student in results}

\item{type}{Filter results to include user with at least one of specified string ("teacher","student","ta","observer","designer")}

\item{role_id}{Integer of role to limit results (SIS ID not supported)}

\item{state}{Filter results to include only users with one of specified string ("active","invited","rejected","completed","inactive")}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get user details for the provided course. Most include parameters ("email",
"enrollments","locked","avatar_url","bio") for this request have all been 
enabled and can be subsetted out if not needed.  This excludes "test_student,"
which has been set to its own argument.
}
