% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiscussionEntries.R
\name{getDiscussionEntries}
\alias{getDiscussionEntries}
\title{Get discussions topic entries from a course}
\usage{
getDiscussionEntries(courseID, topicID, server = "test", ...)
}
\arguments{
\item{courseID}{Course ID to find assignments for}

\item{topicID}{Topic Id to return the entries of}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get discussions for the provided course. Most include parameters ("email",
"enrollments","locked","avatar_url","bio") for this request have all been 
enabled and can be subsetted out if not needed.  This excludes "test_student,"
which has been set to its own argument.
}
