% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiscussionTopic.R
\name{getDiscussionTopic}
\alias{getDiscussionTopic}
\title{Get a single discussions topic from a course}
\usage{
getDiscussionTopic(courseID, topicID = NULL, server = "test", ...)
}
\arguments{
\item{courseID}{Course ID to find discussions in}

\item{topicID}{Topic ID to retrieve}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get discussions for the provided course. Most include parameters for this request have all been 
enabled and can be subsetted out if not needed.
}
