% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiscussionTopics.R
\name{getDiscussionTopics}
\alias{getDiscussionTopics}
\title{Get discussions topics from a course}
\usage{
getDiscussionTopics(url, courseID, search_term = NULL, scope = NULL,
  order_by = "position", only_announcements = NULL, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{courseID}{Course ID to find assignments for}

\item{search_term}{Filter results by term. Must be at least 3 characters.}

\item{scope}{Limit results to a topic state ("locked", "unlocked", "pinned", "unpinned")}

\item{only_announcements}{Set to "true" for only announcements}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get discussions for the provided course. Most include parameters ("email",
"enrollments","locked","avatar_url","bio") for this request have all been 
enabled and can be subsetted out if not needed.  This excludes "test_student,"
which has been set to its own argument.
}
