% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDomains.R
\name{getDomains}
\alias{getDomains}
\title{Get Canvas Domains}
\usage{
getDomains(url, name = NULL, domain = NULL, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation.}

\item{name}{Optional string to retreive only matching results (e.g. "university")}

\item{domain}{Optional string to retreive only matching results (e.g. "edu")}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Gets a list of Canvas Institutions and the base URI for the Canvas.
}
