% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrollments.R
\name{getEnrollments}
\alias{getEnrollments}
\title{Get enrollments for an section, course, or user}
\usage{
getEnrollments(ID, type = "course", server = "test", ...)
}
\arguments{
\item{ID}{ID to find enrollments for}

\item{type}{Use "course", "account", or "user"}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get enrollments for a course, section, or user
}
