% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExternalTools.R
\name{getExternalTools}
\alias{getExternalTools}
\title{Get a list of external tools used ina course}
\usage{
getExternalTools(
  ID,
  type = "course",
  search_term = NULL,
  selectable = NULL,
  include_parents = NULL,
  server = "test",
  ...
)
}
\arguments{
\item{ID}{Account, course or group ID to list external tools in}

\item{type}{Account, course or group}

\item{search_term}{Filter results by partial title. Must be at least 3 characters.}

\item{selectable}{If true, then only tools that are meant to be selectable are returned}

\item{include_parents}{If true, then include tools installed in all accounts above the current context}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of pages in course or group.
}
