% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFiles.R
\name{getFiles}
\alias{getFiles}
\title{Get files}
\usage{
getFiles(ID, IDtype = "course", fileID = NULL, search = NULL,
  content = NULL, quota = FALSE, server = "test", ...)
}
\arguments{
\item{ID}{The course, group, user, folder or specific file ID to return list for.}

\item{IDtype}{The type of ID ("course", "group", "user", "folder", "file")}

\item{fileID}{The ID of a file if getting a specific file.}

\item{content}{Character of file type and or subtype pairs (e.g., 'image/jpeg' or 'image')}

\item{quota}{Boolean of whether this is a request for the file space quota}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}

\item{term}{Character of search term to limit results}
}
\description{
Get a list of available reports
}
