% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFolders.R
\name{getFolders}
\alias{getFolders}
\title{Get folders}
\usage{
getFolders(url, ID, IDtype = "course", folderID = NULL, path = NULL, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{ID}{The course, group, user or folder ID to return a list of folders of.}

\item{IDtype}{The type of ID ("course", "group", "user", or "folder")}

\item{folderID}{The ID of a folder if also providing a course or group ID.}

\item{path}{Character of path to folder}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of available reports
}
