% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMigrations.R
\name{getMigrations}
\alias{getMigrations}
\title{List migrations}
\usage{
getMigrations(ID, type = "course", server = "test", ...)
}
\arguments{
\item{ID}{Integer of the account ID to find sub-accounts for}

\item{type}{String of type - course, account, group, or user}

\item{server}{Test, beta, prodcution, OR alternative name in R.environ OR url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
List content migrations.
}
