% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMigrators.R
\name{getMigrators}
\alias{getMigrators}
\title{List migrators}
\usage{
getMigrators(url, ID, type = "course", ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{ID}{Integer of the account ID to find sub-accounts for}

\item{type}{String of type - course, account, group, or user}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get terms in a specified account.
}
