% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutcome.R
\name{getOutcome}
\alias{getOutcome}
\title{Show the details of an outcome}
\usage{
getOutcome(ID, server = "test", ...)
}
\arguments{
\item{ID}{Outcome ID}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of outcomes groups in a course or account.
}
