% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutcomeGroup.R
\name{getOutcomeGroup}
\alias{getOutcomeGroup}
\title{Show the specified outcome group}
\usage{
getOutcomeGroup(url, ID, groupID, type = "account", ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{ID}{Course or account ID to retreive outcomes group from}

\item{groupID}{Outcome group id to create the outcome in}

\item{type}{Character of "course" or "account"}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of outcomes groups in a course or account.
}
