% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutcomeGroups.R
\name{getOutcomeGroups}
\alias{getOutcomeGroups}
\title{Get outcome groups for a context}
\usage{
getOutcomeGroups(ID, type = "account", server = "test", ...)
}
\arguments{
\item{ID}{Course or account ID to retreive outcomes groups from}

\item{type}{Character of "course" or "account"}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of outcomes groups in a course or account.
}
