% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutcomeLinks.R
\name{getOutcomeLinks}
\alias{getOutcomeLinks}
\title{Get all outcome links for a context}
\usage{
getOutcomeLinks(url, ID, type = "account", ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{ID}{Course or account ID to retreive outcomes groups from}

\item{type}{Character of "course" or "account"}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of outcomes groups in a course or account.
}
