% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutcomeResults.R
\name{getOutcomeResults}
\alias{getOutcomeResults}
\title{Get the results of the outcomes for the students}
\usage{
getOutcomeResults(
  ID,
  userIDs = NULL,
  outcomeIDs = NULL,
  include = c("alignments", "outcomes", "outcomes.alignments", "outcome_groups",
    "outcome_links", "outcome_paths", "users"),
  include_hidden = TRUE,
  server = "test",
  ...
)
}
\arguments{
\item{ID}{Course ID of the outcome}

\item{userIDs}{Vector of user IDs of the students}

\item{outcomeIDs}{Vector of outcome IDs of the outcomes}

\item{include}{Options to include in the results}

\item{include_hidden}{Boolean to include results hidden from learning mastery gradebook}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get the results of the outcomes for the students
}
