% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutcomes.R
\name{getOutcomes}
\alias{getOutcomes}
\title{Show the outcomes for the outcomes group}
\usage{
getOutcomes(ID, groupID, type = "account", server = "test", ...)
}
\arguments{
\item{ID}{Course or account ID to retreive outcomes group from}

\item{groupID}{Outcome group id to get the outcome from}

\item{type}{Character of "course" or "account"}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of outcomes groups in a course or account.
}
