% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPages.R
\name{getPages}
\alias{getPages}
\title{Get pages of a course}
\usage{
getPages(url, ID, sort = NULL, order = NULL, search_term = NULL,
  published = NULL, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{ID}{Course or group ID to retreiv pages of}

\item{sort}{Character of how to sort list ("title","created_at","updated_at")}

\item{order}{Character of order to sort ("asc","desc")}

\item{search_term}{Filter results by partial title. Must be at least 3 characters.}

\item{published}{Boolean to include only published pages}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of pages in course or group.
}
