% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuizSubmissions.R
\name{getQuizSubmissions}
\alias{getQuizSubmissions}
\title{Get submissions for a quiz}
\usage{
getQuizSubmissions(courseID, quizID, server = "test", ...)
}
\arguments{
\item{courseID}{Course ID to find assignments for}

\item{quizID}{Assignmnet ID for the assignment to retreive a submission of}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get the submissions for the provided course and assignment.
}
