% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportStatus.R
\name{getReportStatus}
\alias{getReportStatus}
\title{Get status of a specific report by ID}
\usage{
getReportStatus(accountID, report_type, reportID, server = "test", ...)
}
\arguments{
\item{accountID}{The account ID to retrieve reports from}

\item{report_type}{The report identifier (e.g., outcome_results_csv)}

\item{reportID}{The id of a specific report request}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of available reports, with extra options to select just a report download url.
}
