% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRoles.R
\name{getRoles}
\alias{getRoles}
\title{Get a account roles}
\usage{
getRoles(accountID, server = "test", ...)
}
\arguments{
\item{accountID}{Account ID of page}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of roles for an account
}
