% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRubrics.R
\name{getRubrics}
\alias{getRubrics}
\title{Get a list of rubrics from an account or course}
\usage{
getRubrics(ID, type = "account", server = "test", ...)
}
\arguments{
\item{ID}{Integer of the account ID to find users for}

\item{type}{Course or account}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of rubrics from an account or course
}
