% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSection.R
\name{getSection}
\alias{getSection}
\title{Get info on a specific section}
\usage{
getSection(sectionID, server = "test", ...)
}
\arguments{
\item{sectionID}{Section ID to find assignments for}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get sections for the provided course. All includes enabled: students, avatar_url, enrollments,
total_students, and passback_status.
}
