% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubaccounts.R
\name{getSubaccounts}
\alias{getSubaccounts}
\title{Get Sub-accounts}
\usage{
getSubaccounts(accountID, recursive = FALSE, server = "test", ...)
}
\arguments{
\item{accountID}{Integer of the account ID to find sub-accounts for}

\item{recursive}{Boolean to search returned subaccounts for subaccounts}

\item{server}{The base url of a Canvas installation}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get details of sub-accounts for a provided account.
}
