% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubmisson.R
\name{getSubmission}
\alias{getSubmission}
\title{Get submission for an assignment}
\usage{
getSubmission(courseID, assignID, userID, server = "test", ...)
}
\arguments{
\item{courseID}{Course ID to find assignments for}

\item{assignID}{Assignmnet ID for the assignment to retreive a submission of}

\item{userID}{User ID to retreive assignment submission for}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get the details of a specific submission for the provided course, assignment and user IDs.
}
