% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTerms.R
\name{getTerms}
\alias{getTerms}
\title{Get Terms}
\usage{
getTerms(accountID, state = NULL, server = "test", ...)
}
\arguments{
\item{accountID}{Integer of the account ID to find sub-accounts for}

\item{state}{Optional string "active" or "deleted" to limit request}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get terms in the specified account.
}
