% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUser.R
\name{getUser}
\alias{getUser}
\title{Get details of a specific user}
\usage{
getUser(url, userID, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{userID}{User ID to find assignments for}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get user details for the provided user id.
}
