% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserCourses.R
\name{getUserCourses}
\alias{getUserCourses}
\title{Get courses for a user}
\usage{
getUserCourses(url, userID, state = NULL, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{userID}{The string of a user to request courses of}

\item{state}{String of enrollment state to limit results to (“unpublished”,“available”,“completed”,“deleted”)}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get course details for the account requesting. All include parameters (see API
documentation) for this request have all been enabled and can be subsetted out 
if not needed.
}
