% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserProfile.R
\name{getUserProfile}
\alias{getUserProfile}
\title{Get profile of a specific user}
\usage{
getUserProfile(userID, server = "test", ...)
}
\arguments{
\item{userID}{User ID to profile of}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get user details for the provided user id. NOTE: Returns as a list.
}
