% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadToken.R
\name{loadToken}
\alias{loadToken}
\title{Load API Token}
\usage{
loadToken(env_var_name)
}
\arguments{
\item{env_var_name}{Character of the name of the env variable}
}
\value{
token A character vector of token
}
\description{
Assumes that the Canvas API key is in the .Renviron file.  Default name of variable is "CanvasApiKey."
If absent, looks in the working directory for txt file with token named token.txt.
}
