% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadURL.R
\name{loadURL}
\alias{loadURL}
\title{Load Server URL}
\usage{
loadURL(server)
}
\arguments{
\item{server}{Either the name of the environment variable for the url, or a provided string beginning with 'http'}
}
\value{
url A character vector of of the url
}
\description{
Assumes that the following url variables have been set in the .Renviron folder: CanvasUrlProd, 
CanvasUrlTest, CanvasUrlBeta.
}
