% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSisIDs.R
\name{makeSisIDs}
\alias{makeSisIDs}
\title{Make SIS ID Vector}
\usage{
makeSisIDs(sisids, type = "course")
}
\arguments{
\item{type}{A character vector of the id type to create (account, course, user, term)}

\item{IDs}{The vector of IDs to be modified to work as supplied id type}
}
\description{
Take a vector of integers and return an character vector that prepares those
SIS IDs to work in place of User IDs
}
