% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginationLinks.R
\name{paginationLinks}
\alias{paginationLinks}
\title{Extract Pagination Links}
\usage{
paginationLinks(response)
}
\arguments{
\item{response}{http response object.}
}
\description{
Extract the links from the headers as named objects.
}
