% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startReport.R
\name{startReport}
\alias{startReport}
\title{Start an account report}
\usage{
startReport(
  accountID,
  report_type,
  server = "test",
  enrollment_term_id = NULL,
  include_deleted = NULL,
  course_id = NULL,
  order = NULL,
  users = NULL,
  accounts = NULL,
  terms = NULL,
  courses = NULL,
  sections = NULL,
  enrollments = NULL,
  groups = NULL,
  group_categories = NULL,
  group_memberships = NULL,
  xlist = NULL,
  user_observers = NULL,
  admins = NULL,
  created_by_sis = NULL,
  start_at = NULL,
  end_at = NULL,
  include_enrollment_state = NULL,
  enrollment_state = NULL,
  enrollment_type = NULL,
  minimum = NULL,
  ...
)
}
\arguments{
\item{accountID}{The account ID to retrieve reports from}

\item{report_type}{The report identifier (e.g., outcome_results_csv)}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of available reports, with extra options to select just a report download url.
}
