% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createConversation.R
\name{createConversation}
\alias{createConversation}
\title{Create a new conversation}
\usage{
createConversation(userID, subject = NULL, body = NULL, server = "test",
  ...)
}
\arguments{
\item{userID}{User id of person to message}

\item{subject}{Character of conversation subject}

\item{body}{Character of body content}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Create a single recipient conversation
}
