% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFile.R
\name{createFile}
\alias{createFile}
\title{Upload a course file}
\usage{
createFile(courseID, name = "", filelocal = "", path = "",
  server = "test", ...)
}
\arguments{
\item{courseID}{Course id to create upload the file to}

\item{name}{The name of the file upon upload.}

\item{filelocal}{The path to the file.}

\item{path}{Path of the folder in course files.}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Upload a local file to a course.
}
