% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGradebookColumn.R
\name{createGradebookColumn}
\alias{createGradebookColumn}
\title{Create a custom grade book column}
\usage{
createGradebookColumn(url, ID, title = NULL, position = 1, hidden = FALSE)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{ID}{Course id to create the page in}

\item{title}{String of column title.}

\item{position}{Integer of column position}

\item{hidden}{Boolean of whether the column is hidden or not}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Create a new custom grade book column for the provided course.
}
