% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMigration.R
\name{createMigration}
\alias{createMigration}
\title{Create a content migration.}
\usage{
createMigration(
  ID,
  type = "course",
  migration_type = "course_copy_importer",
  settings,
  server = "test",
  ...
)
}
\arguments{
\item{ID}{ID of context to import into.}

\item{type}{String of type - course, account, group, or user}

\item{migration_type}{String of migrator type}

\item{settings}{Named list of settings for the migrator}

\item{server}{Test, beta, prodcution, OR alternative name in R.environ OR url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Create a content migration with migrator type and a settings list. Does not
currently support file uploads.  Limited testing conducted on various migrators.
}
