% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editGradebookColumnData.R
\name{editGradebookColumnData}
\alias{editGradebookColumnData}
\title{Create/edit a custom grade book column data entry for user in a course}
\usage{
editGradebookColumnData(courseID, columnID, userID, content, server = "test",
  ...)
}
\arguments{
\item{courseID}{Course ID with the custom column}

\item{columnID}{ID of of the column}

\item{userID}{ID of the user (student) to write column datum to}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}

\item{include_hidden}{Character of the content to place in the column for user}
}
\description{
Create/edit a new custom grade book column entry for the provided user and course.
}
