% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccounts.R
\name{getAccounts}
\alias{getAccounts}
\title{Get Accounts}
\usage{
getAccounts(server = "test", ...)
}
\arguments{
\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get details of accounts that you have access to. Parameters 
(lti_guid|registration_settings) for this request have all been enabled and 
can be subsetted out if not needed.
}
