% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnalytics.R
\name{getAnalytics}
\alias{getAnalytics}
\title{Get account or course analytics}
\usage{
getAnalytics(url, ID, type = "activity", term = "current", course = FALSE,
  studentID = NULL, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{ID}{The account or course ID to retrieve analytics for}

\item{type}{The analytics type (e.g., activity, grades, statistics, assignments, communication, student_summaries)}

\item{term}{The term ID, "complete", or "current"(for account analytics only)}

\item{course}{Boolean to specify it the request is for a course.}

\item{studentID}{ID of a student for student specific requests (activity, assignment, communication)}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get account or course analytics
}
