% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCourse.R
\name{getCourse}
\alias{getCourse}
\title{Get a single course}
\usage{
getCourse(url, courseID, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{courseID}{The base url of a Canvas installation}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get details of a single course
}
