% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDomains.R
\name{getDomains}
\alias{getDomains}
\title{Get Canvas Domains}
\usage{
getDomains(name = NULL, domain = NULL, server = "test", ...)
}
\arguments{
\item{name}{Optional string to retreive only matching results (e.g. "university")}

\item{domain}{Optional string to retreive only matching results (e.g. "edu")}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Gets a list of Canvas Institutions and the base URI for the Canvas.
}
