% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGradebookColumnData.R
\name{getGradebookColumnData}
\alias{getGradebookColumnData}
\title{Get custom grade book column data for a course}
\usage{
getGradebookColumnData(url, courseID, columnID, include_hidden = TRUE, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{courseID}{Course ID with the custom column}

\item{columnID}{ID of of the column}

\item{include_hidden}{Boolean to include hidden columns in request}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get custom grade book column entry for the provided user and course.
}
