% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGradebookColumns.R
\name{getGradebookColumns}
\alias{getGradebookColumns}
\title{Get custom gradebook columns from a course}
\usage{
getGradebookColumns(url, courseID, include_hidden = TRUE, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{courseID}{Course ID to gradebook columns}

\item{include_hidden}{Boolean to include hidden columns in request}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get custom gradebook columns for the provided course.
}
