% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutcome.R
\name{getOutcome}
\alias{getOutcome}
\title{Show the details of an outcome}
\usage{
getOutcome(url, ID, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{ID}{Outcome ID}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of outcomes groups in a course or account.
}
