% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutcomeGroup.R
\name{getOutcomeGroup}
\alias{getOutcomeGroup}
\title{Show the specified outcome group}
\usage{
getOutcomeGroup(ID, groupID, type = "account", server = "test", ...)
}
\arguments{
\item{ID}{Course or account ID to retreive outcomes group from}

\item{groupID}{Outcome group id to create the outcome in}

\item{type}{Character of "course" or "account"}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get details of a specified outcomes group. Note: Returns a list.
}
