% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutcomeLinks.R
\name{getOutcomeLinks}
\alias{getOutcomeLinks}
\title{Get all outcome links for a context}
\usage{
getOutcomeLinks(ID, type = "account", server = "test", ...)
}
\arguments{
\item{ID}{Course or account ID to retreive outcomes groups from}

\item{type}{Character of "course" or "account"}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of outcomes in a course or account.
}
