% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutcomeResults.R
\name{getOutcomeResults}
\alias{getOutcomeResults}
\title{Get the results of the outcomes for the students}
\usage{
getOutcomeResults(ID, include = c("alignments", "outcomes",
  "outcomes.alignments", "outcome_groups", "outcome_links", "outcome_paths",
  "users"), include_hidden = TRUE, server = "test", ...)
}
\arguments{
\item{ID}{Course ID of the outcome}

\item{include}{Options to include in the results}

\item{include_hidden}{Boolean to include results hidden from learning mastery gradebook}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}

\item{userIDs}{User IDs of the students}

\item{outcomeIDs}{Outcome IDs of the outcomes}
}
\description{
Get the results of the outcomes for the students
}
