% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPage.R
\name{getPage}
\alias{getPage}
\title{Get a specific page}
\usage{
getPage(url, courseID, pageID, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{courseID}{Course or group ID of page}

\item{pageID}{Page ID number}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get a list of pages in course or group.
}
