% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuizzes.R
\name{getQuizzes}
\alias{getQuizzes}
\title{Get quizzes from a course}
\usage{
getQuizzes(url, courseID, search_term = NULL, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{courseID}{Course ID to find quizzes for}

\item{search_term}{Filter results by partial course name, code, or full ID to match and return in the results list. Must be at least 3 characters.}
}
\description{
Get quizzes for the provided course.
}
