% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSections.R
\name{getSections}
\alias{getSections}
\title{Get sections from a course}
\usage{
getSections(courseID, server = "test", ...)
}
\arguments{
\item{courseID}{Course ID to find assignments for}

\item{server}{Test, beta, production, or other name in R.environ OR full url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get sections for the provided course. All includes enabled: students, avatar_url, enrollments,
total_students, and passback_status.
}
