% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTerms.R
\name{getTerms}
\alias{getTerms}
\title{Get Terms}
\usage{
getTerms(url, accountID, state = NULL, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{accountID}{Integer of the account ID to find sub-accounts for}

\item{state}{Optional string "active" or "deleted" to limit request}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get terms in a specified account.
}
