% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUser.R
\name{getUser}
\alias{getUser}
\title{Get details of a specific user}
\usage{
getUser(userID, server = "test", ...)
}
\arguments{
\item{userID}{User ID to get details for}

\item{server}{Test, beta, prodcution, OR alternative name in R.environ OR url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get user details for the provided user id. NOTE: Returns a list.
}
