% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserCourses.R
\name{getUserCourses}
\alias{getUserCourses}
\title{Get courses for a user}
\usage{
getUserCourses(userID, state = NULL, server = "test", ...)
}
\arguments{
\item{userID}{The string of a user to request courses of}

\item{state}{String of enrollment state to limit results to (unpublished, available, completed, deleted)}

\item{server}{Test, beta, prodcution, OR alternative name in R.environ OR url of server}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get course details for the account requesting. All include parameters (see API
documentation) for this request have all been enabled and can be subsetted out 
if not needed.
}
