% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserPageViews.R
\name{getUserPageViews}
\alias{getUserPageViews}
\title{Get details of a specific user}
\usage{
getUserPageViews(url, userID, start_time = "", end_time = "", ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{userID}{User ID to find assignments for}

\item{start_time}{POSIXct object of date and time of due date}

\item{end_time}{POSIXct object of date and time of due date}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get user details for the provided user id.
}
