% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserProfile.R
\name{getUserProfile}
\alias{getUserProfile}
\title{Get profile of a specific user}
\usage{
getUserProfile(url, userID, ...)
}
\arguments{
\item{url}{The base url of a Canvas installation}

\item{userID}{User ID to profile of}

\item{...}{Optional page options to pass to processRequest}
}
\description{
Get user details for the provided user id.
}
